*UPDATED* 3/26
Changes ArrayList structure to HashMap to allow for fast searching, as well as the year, month, and day views.  Now allows for info to be added and cleared from each day.  Week views need to be implemented, as well as functionality for features such as 'apply to each year' and such.  Still needs to be translated onto another platform.  This should be enough to present.